download_controller/requirements.txt
confluent_kafka==2.3.0
Flask==3.0.0
mysql_connector_python==8.0.26
Werkzeug==3.0.1
prometheus-client==0.19.0
